+++ TeroSubtitler/subtitleapi/UWSubtitleAPI.Formats.EBU.pas
@@ -37,8 +37,8 @@ type
     function IsTextBased: Boolean; override;
     function HasStyleSupport: Boolean; override;
     function IsMine(const SubtitleFile: TUWStringList; const Row: Integer): Boolean; override;
-    function LoadSubtitle(const SubtitleFile: TUWStringList; const FPS: Single; var Subtitles: TUWSubtitles): Boolean; override;
-    function SaveSubtitle(const FileName: String; const FPS: Single; const Encoding: TEncoding; const Subtitles: TUWSubtitles; const SubtitleMode: TSubtitleMode; const FromItem: Integer = -1; const ToItem: Integer = -1): Boolean; override;
+    function LoadSubtitle(const SubtitleFile: TUWStringList; const FPS: Double; var Subtitles: TUWSubtitles): Boolean; override;
+    function SaveSubtitle(const FileName: String; const FPS: Double; const Encoding: TEncoding; const Subtitles: TUWSubtitles; const SubtitleMode: TSubtitleMode; const FromItem: Integer = -1; const ToItem: Integer = -1): Boolean; override;
   end;
 
 // -----------------------------------------------------------------------------
@@ -74,7 +74,7 @@ end;
 
 // -----------------------------------------------------------------------------
 
-procedure TimeCodeToByteArray(var Dest: array of Byte; const Source: Cardinal; const FPS: Single);
+procedure TimeCodeToByteArray(var Dest: array of Byte; const Source: Cardinal; const FPS: Double);
 var
   s: String;
 begin
@@ -139,7 +139,7 @@ end;
 
 // -----------------------------------------------------------------------------
 
-function TUWEBU.LoadSubtitle(const SubtitleFile: TUWStringList; const FPS: Single; var Subtitles: TUWSubtitles): Boolean;
+function TUWEBU.LoadSubtitle(const SubtitleFile: TUWStringList; const FPS: Double; var Subtitles: TUWSubtitles): Boolean;
 var
   UWFile      : TUWFile;
   GSIBlock    : TGSIBlock;
@@ -213,7 +213,7 @@ end;
 
 // -----------------------------------------------------------------------------
 
-function TUWEBU.SaveSubtitle(const FileName: String; const FPS: Single; const Encoding: TEncoding; const Subtitles: TUWSubtitles; const SubtitleMode: TSubtitleMode; const FromItem: Integer = -1; const ToItem: Integer = -1): Boolean;
+function TUWEBU.SaveSubtitle(const FileName: String; const FPS: Double; const Encoding: TEncoding; const Subtitles: TUWSubtitles; const SubtitleMode: TSubtitleMode; const FromItem: Integer = -1; const ToItem: Integer = -1): Boolean;
 var
   Stream   : TStream;
   Buffer   : TBytes;
